fix: 修改時間格式

+++ app/src/main/java/me/lazy_assedninja/android_dumpling_timer/data/db/Setting.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class Setting(
     @PrimaryKey val id: Int,
-    @ColumnInfo(name = "base_time") val baseTime: Int,
-    @ColumnInfo(name = "gap_time") val gapTime: Int,
-    @ColumnInfo(name = "sound_effect_time") val soundEffectTime: Int
+    @ColumnInfo(name = "base_time") val baseTime: Long,
+    @ColumnInfo(name = "gap_time") val gapTime: Long,
+    @ColumnInfo(name = "sound_effect_time") val soundEffectTime: Long
 )
