+++ app/src/main/java/me/lazy_assedninja/android_dumpling_timer/data/repository/SettingRepository.kt
@@ -7,7 +7,7 @@ import me.lazy_assedninja.android_dumpling_timer.data.db.SettingDao
 
 class SettingRepository(private val context: Context, private val dao: SettingDao = AppDatabase.getInstance(context).settingDao()) {
 
-    fun setSetting(baseTime: Int, gapTime: Int, soundEffectTime: Int) {
+    fun setSetting(baseTime: Long, gapTime: Long, soundEffectTime: Long) {
         dao.insert(Setting(0, baseTime, gapTime, soundEffectTime))
     }
